Turmas do Território do Saber não está atualizando os dados

+++ src/SME.SGP.Infra/Dtos/DisciplinaTotalDto.cs
@@ -4,7 +4,7 @@ namespace SME.SGP.Infra
 {
     public class DisciplinaTotalDto
     {
-        public int DisciplinaId { get; set; }
+        public long DisciplinaId { get; set; }
         public int Total { get; set; }
     }
 }
